All functionalities have been implemented. Time to merge them.

+++ src/rest_api/database/database.py
@@ -28,6 +28,9 @@ class Database:
             except OSError as exc:  # Guard against race condition
                 if exc.errno != errno.EEXIST:
                     raise
-        db_file = open(db_file_name, 'w')
+        if not os.path.isfile(db_file_name):
+            db_file = open(db_file_name, 'w')
+        else:
+            db_file = open(db_file_name, 'r')
         db_file.close()
         return sqlite3.connect(db_file_name, check_same_thread=False)
